Changed Rate to $40/day for Coastal Quarters
Updated rate to $40 per night.

Removed the gears for the settings page under the admin screen.
Fixed the NC State User image from appropriately showing.
Updated the linked PDF document on the front page of making a reservation for Coastal Quarters. The PDF that we are using is the same from Avent Ferry GuestRoom reservations.