Updated Timezone, cleaned up confirmation message
Changed Timezone from UTC to America/NY -- added Arial font to the
confirmation message that is sent to the billing user.

Finished up the create a new reservation page by asking for OUC number,
project grant number and bookkeeper information when needed. Prefilled
out these fields with N/A when the end user selects "Guest" from the
radio button.

Removed the requirement indicator for more information as it is no
longer required in order to complete the reservation.