Removal of the extra atStop on views causing page not to load.
Removed the extraneous @stops that are no longer necessary with the newer version of Laravel. Having these in the two blades caused the page not to render. There could be a few other pages with this issue; but given the limited functionality of this web site; this should be all instances.

Added a span to correctly display the "List Users" link when the left sidebar is collapsed.